Obsolete with error old transport settings extensions
